+++ Assets/Scripts/Enemies/SO/ShootAndSpawnEnemyScriptableObject.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class ShootAndSpawnEnemyScriptableObject : ScriptableObject
 {
 	public float speed;
-	public int damage;
-	public int health;
+	public float damage;
+	public float health;
 
 	public bool isShield;
 	public float distanceFromMotherShipToStopAndShoot;
